+++ zerver/lib/digest.py
@@ -10,8 +10,9 @@ from django.db.models import Q, QuerySet
 from django.template import loader
 from django.conf import settings
 
+from zerver.lib.send_email import send_future_email
 from zerver.lib.notifications import build_message_list, hash_util_encode, \
-    send_future_email, one_click_unsubscribe_link
+    one_click_unsubscribe_link
 from zerver.models import UserProfile, UserMessage, Recipient, Stream, \
     Subscription, get_active_streams
 from zerver.context_processors import common_context
